feat: add material-ui-react-kit, switch to reponsive menu
Add material-ui styles from https://demos.creative-tim.com/material-kit-react/#/

Stitch the menu over to a responsive one where the menu is fixed an optn for
the desktop version and hidden behind a button for the mobile version.

Covert the Banner image over to a SVG with a dynamically set date sting.