fix(ui): constrain content width
The current welcome page dumps out some long text that included the url for the
gravatar image, this is long and hard to wrap. This in turn screws up the
overall width of the page offsetting the rest of the disply, especially on
mobile. Setting this top level width fixes things.