feat(dev): use docker for all local databases
I use both postgres and mysql to handle migrating data from the existing mysql v1 database, to the new postgres database. This change moves them form local installations to servers running docker.  The biggest change was refactoring the import scripts to better parameterize the mysql connection information.

Also default server and client ports in dev are moved up to 30001 and 30000 so that they are well clear of any other development I might be doing.  The same sort of shift happens for the docker containers.