feat(security): update libs to pass yarn audit
Note that this involves using the forced resolution for exec-sh/merge that is used by most of jest etc.  Since the jest version is fixed by create-react-app, this is the easiest way to avoid the problem and make sure that we don't depend on a vulnerable version.

Most other libs (that aren't pinned by create-react-app were also brought up to date).