feat(graphql): stop hacking query params and use client gql for url
The scrolling behavior around the PastCons page worked by effectively storing a flag in the query params related to the the source of the url change, either via an explicit set coming from the menu (which will require a jump), or from the scrolling behavior which requires no further action.

Now the source of the url is stored in the client side gql cache, cleaning up the url.