update user registration
Extend the api to better handle authentication.

*  refactor the api to avoid duplicating so much code.
*  support querying auth0
  *  to validate passwords
  *  to send password reset links
  *  unused - to get an api token for the management api

Change login to only login if the email is verified, displaying a message if the email isn't verified.

Added notistack so that we can have stacked toast messages.

Allow editing the user profile.  Note that you can't change the email address yet.

Allow changing the password if this is an auth0 account.  Note that this simply causes auth0 to email out a change password link.

Fixed auth0 to create a user if one was not found in the system for a given email address.  Note that it only does this for verified email addresses.