Error cleanup
Changed my pattern for handling graphql errors to check before looking at the loading or the data indicators since in places I check for !!data as well as loading to display the loading indicator, this in turn masks errors and just shows the loading indicator in perpetuity.  That's a bad pattern, so I moved all of the error testing first where it's always going to take precedence since it's less likely to get modified to be broken that way.