More Game Selection stuff
Submitting choices works,  still need to finish off the final submission part.

Caveats, the screen refresh is accurate but a bit slow at times.  If there's time, should re-work it.

Fixed a couple of bugs caused by forgetting that zero was a valid value, causing me to update the eslint rules, adding '@typescript-eslint/prefer-nullish-coalescing', and also change a lot of uses of `||` to `??`.

Added a standard decorator pattern to many of the game book components, allowing the same infrastructure to be used in all places that display an sort of game book.

Externally, changed the vercel project configuration to use project environment variables, as such we no longer need the now.json.

Updated various libs.