Lots of game choice changes.
Switch from local storage to session storage.  I'd like to switch back at some point, but I need a better way to deal with stale data in the cache and don't have the time to deal with it right now.

Really fix the game book and game signup routes to not barf if the user is going to the site in a new session - can't have it deciding that they aren't a member before the data has loaded and giving them an error.

 Add a goto Top button to the game signup page.

 Add a database stored function to create the initial game choices - this makes setting the gm information much easier and faster.

 End confirmation email for game choices.