Roll back the paginated filtered user dropdown
The Dropdown worked fine, the problem is that it changes the global users definition in apollo requiring that every access to the users object also be paginated.  That's an issue that I just don't want to deal with right now - I'm not even convinced that I need to worry about that issue in this app given the data sizes.

Also, there are many things that are quite simply easier to do on the front end where I have access to all of the relevant data and so for as long as I can manage keeping that working, the better.  Table search is a good example - it's trivial to do on the client side, but I'd have to add full text search on the backend and then add the appropriate postgraphile plugin to expose it to get the same effect - certainly worth it if the data set gets too large, I'm just not convinced that we'll ever run in to that problem.

We're certainly not there right now.

I also find that the more that I use Apollo, the more that I begin to dislike some of their choices.  This is one of them.