Remove MaterialKit code
Most of the MaterialKit code wasn't actually being used so it was silly
to keep it around.  That said, the styles are still heavily based on
that MaterialKit, they are just migrated over to the theme object.

This made it much easier to refactor all of the component tree out of
components/Acnw into plain old components.