Clean up new account UI (#23)
Tidy up code around first login.  The split between logged in and not
logged in wasn't really accurate. There are a couple of less obvious
states there that only show up during the initial registration for the
convention. Fist when the email hasn't been verified, and then after
verification, before the user account has been created - and that
account is only created by a successful login (via rule on auth0).

We only create users for verified accounts.  On one hand that causes
this problem, on the other it cuts down on fake account immensely.