Enable next.js modular imports (#44)
* Enable next.js modularizeImports

This is an optimization that (like the babel equivalent) maps imports of the form `import { Box } from '@mui/material'` into `import Box from '@mui/material/Box'`.  The former is nicer for the dev, the latter is better when it comes to tree-shaking.  It has some benefits for `@mui/material`, but far better benefits for `@mui/icons/material` where the full import is in excess of 2k icons.

The config for this is in the next.config.js files.  Ideally that would have bene all of it, sadly `@mui/material` is a bit inconsistent about its internal paths, and `useTheme` isn't in `@mui/material/useTheme` but is instead in `@mui/material/styles` :(  The majority of the changes are to pull that in explicitly so that the plugin works correctly.

There are a couple of unrelated changes, but they aren't significant enough to warrant separate PRs.

* Dev readme added for acus, it's not perfect, but it's a start.
* The display of local environment flags was broken, apparently I refactored it a while ago and hadn't tested it.
* The logging of email responses should have been dev only, now it is.

* fix lint errors