Add global table action to download content as CSV
Action added in TableToolbar. Takes the visible columns and formats
them as a CSV row, then iterates through the rows, rendering each,
necessary for the not-shown pages, especially if you change columns
shown/hidden, and formats non-numerics as strings, enclosing in
double quotes and escaping double quotes, and concatenating them
as rows of CSV, separated by a newline.

It's all thrown in a Blob, a hidden link created with the url set
to an object URL to the Blob, the link clicked, and then destroyed.

Resolves: #54