Improve client side route protection
Previously certain routes only showed up in the nav if you had
permission, but if you knew their path, you could navigate to them and
then deal with the backend errors that rippled up to the UI.

With this change, those paths all have the same permissions checked
before the component is rendered.  This is closer to how this worked
before the next.js migration, but the routing changes let this slip
through.