Move config into database
Migrate the existing static config over to database fields.  This is
done by mapping config fields to a zod validated schema, there the
dotted name convention maps to chunks in a hierarchical object.

The config now contains sections for flags, config, and urls.

The main "trick" to making this work effectively is by using the next.js
approach to loading props on the server and passing them down to the
client, paired with pre-rendering the pages on the server.

Lots of renames to make this look clean, as well as functions to access
what used to be globals.

The biggest code pain is the littering of pages  with `export const
getServerSideProps = configGetServerSideProps` statements.