Config and Setting updates (#125)
Several changes

*  All config values have been moved into settings and loaded from the database. This is the bulk of the ripple.
*  Due to the intertwined nature of the settings, a new convention year can be kicked off via new action on the settings page that automatically increases the convention year, prompts for new con dates and then updates a batch of other values accordingly. Deadline dates move forward a year, flags get switched off, that sort of thing.
*  The settings page has tabs for each of the major setting hives, this splits up an otherwise unwieldy list.
api usage has a new direct postgraphile pattern that avoids the api calling back to itself, that's inline with next.js best practices.
*  Naming for config object, flags etc have been cleaned up somewhat. There's a reasonable schema validation on expected values.
*  Settings have been updated to allow for date and numeric values.
*  a migrator has been written based on the old setting code to dump the historic values into the database.
* update DatePicker
* library updates

* Move config into database

-- other commit notes

Migrate the existing static config over to database fields.  This is
done by mapping config fields to a zod validated schema, there the
dotted name convention maps to chunks in a hierarchical object.

The config now contains sections for flags, config, and urls.

The main "trick" to making this work effectively is by using the next.js
approach to loading props on the server and passing them down to the
client, paired with pre-rendering the pages on the server.

Lots of renames to make this look clean, as well as functions to access
what used to be globals.

The biggest code pain is the littering of pages  with `export const
getServerSideProps = configGetServerSideProps` statements.