payment confirmation email
This is terribly basic, but it works and does tell the user what they
want to see, albeit in a rather stark manner.  Frankly the whole email
templating system needs a big refactor.  But that's a problem for
another day.

The actual email bit is pretty small, bu I cleaned up some naming to
make things a bit more consistent and changed where the configuration
was coded so that it could be better shared.