Stripe library Upgrade
Note that this only upgrades the Stripe loader, that loader always loads the latest stripe.js file.

Also update the version of the Stripe API that we use, this was synchronized with a change to the ACNW Stripe api on the stripe site.

Somewhat related, renamed the column user.amountOwed -> user.balance since that was how it was being used, and the negative amount in the amountOwed field kept confusing me.