Fix date config error
As reported by a number of folk on the east coast, convention dates appeared to be incorrect.  Looking in the database it was clear that they weren't being created correctly and that in turn was causing them to appear incorrectly for a number of folk looking at them from non west coast time zones.

This change creates them correctly, i.e. convention dates should be in the tz of the convention.

Then the configuration is fixed up to create the dates in the correct tz.