Remove GM Preview of Player Lists from the timeline
Never got much good out of it, so at ACUS we'll experiment with not
having the GMs preview their player lists.

Really should check if the date value exists or has a magical value
(NULL, 1/1/1970, ...) and not display the section, or just iterate
over a list of database dates and generate the repeating LI section
if it's a valid date. But that'll be for a later date.

Resolves: #151