Remove GM Player Lists preview
Remove player list preview date for GMs. At ACUS we haven't seen much
good feedback, so we are going to try to do without this year.

Really might want to be a function that takes the label and a database
field and generates it if the database field has a valid date in the
future, maybe display it different if past, maybe leave it out if it
is missing, has a magic value, or is null.

That will be another day.

Resolves: #151