Take on ACNW's Stripe changes and cosmetic diffs (#150)
* Take on ACNW's Stripe changes and cosmetic diffs

Copied in the ACNW payment changes, made matching changes
to Membership Wizard and many minor differences that were
mostly cosmetic, just to make big diffs easier.

Also adopted the donate to help out (though haven't made
the changes to pick up them up in membership / signup).

Resolves: #144, #148

* Add frozen lockfile for PNPM as Vercel complained

* Remove GM Preview of Player Lists from timeline

Remove HTML and display. Really should check if value exists or has a
magical value not to display the section.

Resolves: #151

* Undo last commit as it should be on another branch

(sigh) Created new branch, didn't switch to it.

Resolves: #151

* Remove GM Preview of Player Lists from the timeline

Never got much good out of it, so at ACUS we'll experiment with not
having the GMs preview their player lists.

Really should check if the date value exists or has a magical value
(NULL, 1/1/1970, ...) and not display the section, or just iterate
over a list of database dates and generate the repeating LI section
if it's a valid date. But that'll be for a later date.

Resolves: #151