Make timeline and deadline dates fully data driven
Changed the Settings to add a second exclusion so the main config page
can ignore the timeline dates (they begin with 'config.timelineDates.')
and to add a tab for Timeline Dates filtering on the same. The config
values are meant all to be dates. As code and value are both 100
characters and the database doesn't care I use the code as the label to
display. Finally added an entry to the configuration to read the
config.timelineDates values and add an array of label and date entries.

Created a CutoffDateWarning component that takes a date key (i.e. the
label) and if the date value of that is in the past it shows a warning
not to register without first contacting registration by email to the
email in ContactEmail configuration. If that date is in the future it
shows nothing.

Created a TimelineList component that takes the array of lable/date
values, and an optional ignore before date. Again, it's a label of any
of the dates. It takes the array, filters out any dates before the date
with that label, sorts it in ascending order on the dates and displays
them. If the date is in the past it uses the class deadlineExpired,
otherwise deadline.

Changed the class deadlineExpired from red (error.main) to a 700 gray
with strike through.

In use it cleans up the Welcome page.

ACUS' is now:

    <Page title='Welcome' titleElement={titleElement}>
      <MdxWithExternalLinks>
        <WelcomeContent />
      </MdxWithExternalLinks>
      <DynamicMemberContent />
      <CutoffDateWarning cutoffDateConfig='Membership Payment In Full Due' />
      <TimelineList ignoreBeforeDateConfig='Registration Open' />
    </Page>

More imortantly, the cutoff date warning doesn't require reading and
interpreting, but shows as needed.

And the timeline dates are easily maintained in settings and ones can
be added/removed, or, for this year, left out, by leaving their value
set to last year's.

Resolves: #166, #167