Various easy dependency updates
Note that I'm avoiding the more impactful updated until I've dealt with
removing graphql and replaced it with trpc.  In large part because the
graphql dependencies are old but also blocked on postgraphile which has
a ton of upgrade issues.  Then next.js breaks on upgrade due to
postgraphile issues.