Switch to unified DATABASE_URL
Use the t3-oss stack env package, that validates expected/required environment variables.

This is a rather large ripple, we use env vars all over the place, but it's noticeably simpler to just have the one variable and parse it when needed as opposed to passing through all of them and recreating the single url whenever you want to access the database.