Tons more migrations from graphql over to trpc
This migrates, user, membership, lookups and lookupValues.

Also, migrates various Decimals in the database over to being floats.
Mostly because Prisma has awkward Decimal support.  Not the best
solution, but I spent a while trying to make the Decimal stuff work and
didn't get anything that I was happy with.

There are small changes to the shape of the returned values which led to
some code changes, but the largest ripple comes from type updates.

Final note, dropping postgraphile for various tables means that the
synthetic nodeId field no longer exists and so there are associated
changes to switch to using the table id column instead - honestly, this
would have been better in the first place.