tRPC refactor phase 1 (#209)
For a variety of reasons I want to migrate away from using graphql in this app, primarily because it (postgraphile) has been slow to upgrade and is tying us to a lot of rather old library versions, because it doesn't play well with recent versions of next.js (blocking those upgrades), and because overall it's a bit heavy handed to run in a serverless environment. Granted, if I could upgrade it and make it play well with next.js I'd probably have felt less urgency to upgrade it, but it doesn't...

Given that, I'm switching over to using tRPC for the transport layer. It's light-weight, well supported in the environments that I want to target, and also offers full type safely between the front and backends. For the backend I'm using Prisma, this is a typesafe ORM that also supports direct usage of raw sql.

This first step migrates, user, membership, settings, lookups, hotelRooms, and hotelRoomDetails over to tRPC, this validates the approach, and shows that both systems can coexist until such time as the other types can be migrated over. As a nice bonus, these new api calls are about 8x faster (likely due to requiring far less cold start time in the serverless environment - no short lived postgraphile server startup costs).

The usage of the old and the new api are very similar, since both interfaces used @tanstack/query for their state and cache management. So the actual calls change syntax, but the rest of the behavior is very similar. The resultant types have also changed, becoming somewhat flatter than the graphql ones. Queries have been changed to stop referring to the graphql nodeId and instead use id.

Related ripples come from generalizing the transport error handling from the old GraphQLError over to the new TransportError.

Note that the change to Prisma requires a single DATABASE_URL variable rather than the various sections as separate vars.

Other changes:

deleted graphiql
renamed client package to graphql-client and create a new client package.
switch to eslint v9 and its flat config, which in turn has a couple of new rules, which caused ripples.
prefer using the nullish coalescing operator ?? instead of using ||, it's always a safer default unless you have a good reason to not use it.
fix a lot of quotes in blocks of text to use html entities.
start using the debug library for runtime logging. This is off by default but can be enabled in the browser console by setting localStorage.debug = '*'.
switched tests over to using vitest
TODO:

Finish off migrating the remaining objects, game, gameAssignments, gameChoices, stripe, and transaction`. That should be the end of the client side graphql stuff. Then pull out the last few bits of server side graphql infra.
refactor the various client and server packages to be somewhat cleaner.

--

* Drop Graphiql, refactor some graphql dependencies

Rename GraphQlError to TransportError and move it out of ui and into amber

* Add Vitest

get tests working

* Switch to unified DATABASE_URL

Use the t3-oss stack env package, that validates expected/required environment variables.

This is a rather large ripple, we use env vars all over the place, but it's noticeably simpler to just have the one variable and parse it when needed as opposed to passing through all of them and recreating the single url whenever you want to access the database.

* Initial pass with trpc working for config and part of settings

* Switch to eslint v9

Which in turn led to some cleanup.

* Settings/Config ported

also the legacy client is now moved to client-graphql

* Tons more migrations from graphql over to trpc

This migrates, user, membership, lookups and lookupValues.

Also, migrates various Decimals in the database over to being floats.
Mostly because Prisma has awkward Decimal support.  Not the best
solution, but I spent a while trying to make the Decimal stuff work and
didn't get anything that I was happy with.

There are small changes to the shape of the returned values which led to
some code changes, but the largest ripple comes from type updates.

Final note, dropping postgraphile for various tables means that the
synthetic nodeId field no longer exists and so there are associated
changes to switch to using the table id column instead - honestly, this
would have been better in the first place.

* hack pnpm version to bypass vercel bug

* fix cert and env validation

* fix data format error when dealing with email interface

* apparently optional urls break

* try and fix prisma so it'll install the rhel driver

* add some debugging to the build

* add some debugging to the build

* try another next and prisma config option

* try another approach to get the cert availabe to prisma on vercel

* remove generated code from repo

* dep updates

* apply prisma changes to acus

* remove last graphql hotelRoom references