Upgrade to TanStack Table v8
By far the largest set of changes here are the ones related to upgrading
from react-table@7 to @tanstack/react-table@8. This required changes in
a large number of files across many packages, as the API has changed
significantly. Key changes include:

- The way columns are defined and configured has been revamped.
- The table instance API has been simplified and made more consistent.
- Several hooks and utility functions have been updated or replaced.

There is also an eslint upgrade that enforced rules about importing
types explicitly. followed by change to try out oxlint and oxfmt instead
of eslint and prettier.

There is a new ui-test app, which is a playground for testing the new
table components in isolation.

There are additional changes to fix the behavior of trpc and server-side
rendering in Next.js apps, ensuring that prefetching can be correctly
marshalled when using superjson.